requirements.txt
pandas>=1.5.0
numpy>=1.23.0
yfinance>=0.2.18
ccxt>=4.0.0
ta>=0.10.2
scikit-learn>=1.2.0
joblib>=1.2.0
fastapi>=0.95.0
uvicorn>=0.21.0
sqlalchemy>=2.0.0,<3.0.0
python-dotenv>=1.0.0
requests>=2.28.0
pytest>=7.0.0
httpx>=0.24.0
pydantic>=2.0.0
prometheus-client>=0.16.0
python-multipart>=0.0.6
pydantic-settings>=2.0.0
redis>=4.5.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
alpaca-trade-api>=3.0.2
